Asset downloader internally uses an array of selected items
Problem: The asset downloader maintains the selection through a variable
that is either None or a single-item. We want to allow for multiple
selection in the future and this mechanism won't work.

Solution: Change the mechanism used to track the selected items to use
an array. The 'None' value is replaced by a check for an empty array.
When needed we use the first item of the array to replace the
single-valued variable. When the other APIs are updated to work with
multiple elements we'll change the code to use the full list of selected
items instead of just the 'head'.