Update asset downloader selection callback to handle multiple rows
Problem: We want to enable multiple row selection in the asset
downloader but most of the code is hardcoded on the assumption that only
a single row of data can be selected.

Solution: Update the callback called when the selection change to
correctly update the state of the selection with possibly multiple
remote assets selected. Update the code that displays the remote assets
to show a summary when more than one asset is selected.

Note: The other parts of the code are already updated to deal with an
array of selected remote assets. They just ignore all assets after the
first. This is not a problem since we didn't change the selection mode
yet and only a single item can be selected.