Enable extended selection in asset downloader
Problem: We want to download multiple assets with a single click.

Solution: Enable the ExtendedSelection on the asset downloader's table
view. In the 'Download' button callback, calls a new recursive method
that will download each asset one by one while displaying the progress.
A new progressbar is added to the 'Selected' box, as well as a new
'Cancel downloads' button. When the cancel buttons is pressed it sets a
flag that indicate to the recursive method that it should not process
the next remote asset.