Fix status bar not updating when slider's edit value is modified
Problem: When modifying a slider through its 'edit' textbox the status
bar is not updated. This is mostly visible when modifying the macro
parameters.

Solution: Uncomment the call to _onChange() from the onChange method
that is called when the slider's value is edited. _onChange() is already
called when the slider is moved.